Set access details properly when creating queues
The queue creation process goes through ...Base->new which expects the first two
parameters to be the access key and the secret key, but it had been receiving a
hash which meant the first argument was a hash key, with the second being the
corresponding value - if the first key was 'Version' or similar, this was ok,
because the access details later in the hash would properly stomp it, but if
the access key was first up it would set the AWSAccessKeyId field in the new
queue to the string 'AWSAccessKeyId' with no ability to stomp it later.
Randomisation of the hash keys means that this was a somewhat difficult to
reproduce problem (and it's possible that the more consistent randomisation
system in earlier versions of perl would never actually have the access key as
the first value).